+++ azalea-client/tests/mine_block_timing_hand.rs
@@ -18,7 +18,7 @@ use azalea_protocol::{
 use azalea_registry::Block;
 
 #[test]
-fn test_mine_block_timing() {
+fn test_mine_block_timing_hand() {
     init_tracing();
 
     let mut simulation = Simulation::new(ConnectionProtocol::Game);