+++ azalea-core/src/registry_holder/block_state_provider.rs
@@ -1,0 +1,3 @@
+/// TODO: unimplemented
+#[derive(Debug, Clone, simdnbt::Deserialize)]
+pub struct BlockStateProvider {}