+++ azalea-core/src/sound.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 
 use crate::identifier::Identifier;
 
-#[derive(Clone, Debug, PartialEq, AzBuf, Serialize)]
+#[derive(Clone, Debug, PartialEq, AzBuf, Serialize, simdnbt::Deserialize)]
 pub struct CustomSound {
     pub sound_id: Identifier,
     pub range: Option<f32>,